Cleaning up commmands
