+++ lib/local/index.js
@@ -15,8 +15,12 @@ var _ = require('lodash'),
     wsEvents = eventsConf.WS.LOCAL,
     fsEvents = eventsConf.FS.LOCAL;
 
-function start(projects, config) {
+function start(projects, opts, config) {
     _.each(projects, function(project) {
+        if (_.isEmpty(config.projects[project])) {
+            console.log(text.PROJECT_NOT_FOUND, project);
+            return false;
+        }
         startProject(project, config);
     });
 }
@@ -88,6 +92,24 @@ function startProject (project, config) {
     });
 }
 
+function once(projects, opts, config) {
+    _.each(projects, function(project) {
+        if (_.isEmpty(config.projects[project])) {
+            return console.log(text.PROJECT_NOT_FOUND, project);
+        }
+        var projectConf = config.projects[project];
+        var log = util.generateLog(projectConf.project, projectConf.hostname);
+        
+        log(text.SYNC_ON_ONCE);
+
+        bigSync(projectConf, {
+            dry: opts.dryRun,
+            debug: config.debug
+        }, _.partial(log, text.SYNC_ON_ONCE_DONE));
+    });
+}
+
 module.exports = {
-    start: start
+    start: start,
+    once: once
 };
